Move View search to EditDomain
To make view searching more independent from the currently used tool it
is moved to an own FBNetworkEditDomain.
